requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.2.2
pydantic==2.12.5
httpx==0.28.1
python-multipart==0.0.22
bcrypt==4.1.3
PyJWT==2.12.1
